Reuse port acquisition logic from background_devnet
