+++ tests/integration/common/mod.rs
@@ -1,6 +1,7 @@
 #![cfg(test)]
 pub mod background_anvil;
 pub mod background_devnet;
+mod background_server;
 pub mod constants;
 pub mod errors;
 pub mod fees;
